requirements.txt
openaigoogle-cloud-visionpillowpyttsx3SpeechRecognitionpytesseractpyautoguirequeststtkbootstrap